# cmd/go-server/go/model_api_response.go
/*
 * Goserver in Open API
 *
 * server template in GO.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

// ApiResponse - A Common Successful API response
type ApiResponse struct {

	Code int32 `json:"code,omitempty"`

	Type string `json:"type,omitempty"`

	Message string `json:"message,omitempty"`
}